requirements.txt
# Python MQTT Simulator Dependencies
paho-mqtt==1.6.1
